Contact
Scrapped the idea of the contact page. Linked in, github, and email will be displayed on the homepage. Also, email will be displayed in the navigation menu at the very bottom. Figured there was no need for an entire page to be devoted to 3 links. Furthermore, need to revamp the About div. Change it to School, Schooling or Classes. Have it display my current semester and semesters to come. Overall, decent progress. Will work more tonight. Contact status: in progress.